Split the navbar into Wrapper, OrganizationMenu and PrimaryMenu
Wrapper owns the nav shell, the logo and the fragment cache, and renders the
other two. logo_only: true renders just the logo, so the doorkeeper layout is
Wrapper::Component.new(logo_only: true) and the .logo_only factory goes away.

OrganizationMenu's render? carries the condition the layout's if/elsif used to
encode, so the signup link is now a plain `if current_user.blank?` beside it.

Translations split per component, which needs a write rule in i18n-tasks.yml —
without one, normalize consolidates both sidecars back into the parent
namespace's file. sign_up is under both scopes; the wrapper's centre link and
the menu item both use it.
