--- app/components/page_block/navbar/component.en.yml
@@ -1,22 +1,0 @@
----
-en:
-  components:
-    page_block:
-      navbar:
-        blog: Blog
-        donate: Donate
-        help: Help
-        log_in: log in
-        logout: Logout
-        marketplace: Marketplace
-        marketplace_messages: Marketplace messages
-        menu: Menu
-        register_a_new_bike: Register a new bike
-        search: Search
-        settings_menu: Settings
-        sign_up: Sign up
-        stolen_bike: Stolen bike?
-        the_nonprofit_bike_registry: the non-profit bike registry
-        user_settings: "%{user_email} settings"
-        view_org: View %{org_name}
-        your_registrations: Your registrations
