+++ app/components/page_block/navbar/primary_menu/component.html.erb
@@ -1,0 +1,73 @@
+<ul id="primary-main-menu" class="primary-main-menu tw:m-0 tw:list-none">
+  <% menu_items.each do |item| %>
+    <% case item[:type] %>
+    <% when :divider %>
+      <li class="divider-nav-item <%= item[:item_class] %>" aria-hidden="true"></li>
+    <% when :settings %>
+      <li class="primary-nav-item">
+        <a
+          href="#"
+          class="nav-link settings-link primary-nav-link-with-submenu"
+          id="setting_submenu"
+          aria-label="<%= translation(".settings_menu") %>"
+          aria-haspopup="true"
+          aria-expanded="false"
+          data-toggle="dropdown"
+        >
+          <svg
+            id="settings-icon"
+            data-name="Layer 13"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 30.16 30"
+          >
+            <defs>
+              <style>.cls-1{fill:none;stroke:#a4a4a4;stroke-miterlimit:10;}</style>
+            </defs>
+
+            <title>settings</title>
+
+            <circle
+              class="cls-1"
+              cx="10.68"
+              cy="19.41"
+              r="3.78"
+            />
+
+            <circle
+              class="cls-1"
+              cx="23.92"
+              cy="6.17"
+              r="1.89"
+            />
+
+            <path
+              class="cls-1"
+              d="M4900.95,1506.07a8.21,8.21,0,0,0-.18-1.71l2-1.15-2.52-4.37-1.95,1.15a6.91,6.91,0,0,0-2.92-1.72v-2.14h-5v2.14a7.28,7.28,0,0,0-3,1.72l-2-1.15-2.53,4.37,2,1.15a8.15,8.15,0,0,0,0,3.43l-2,1.14,2.52,4.37,2-1.15a7.26,7.26,0,0,0,3,1.72v2.29h5v-2.29a6.93,6.93,0,0,0,2.92-1.72l2,1.15,2.51-4.37-2-1.14A8.28,8.28,0,0,0,4900.95,1506.07Z"
+              transform="translate(-4882.07 -1486.66)"
+            />
+
+            <path
+              class="cls-1"
+              d="M4910.4,1492.83a4.42,4.42,0,0,0-.12-1l1.25-.72-1.26-2.18-1.21.73a4.53,4.53,0,0,0-1.72-1v-1.45h-2.52v1.45a4.86,4.86,0,0,0-1.81,1l-1.28-.73-1.27,2.18,1.25,0.72a4.27,4.27,0,0,0,0,2l-1.26.72,1.26,2.19,1.3-.73a4.84,4.84,0,0,0,1.81,1v1.45h2.52v-1.45a4.51,4.51,0,0,0,1.72-1l1.23,0.73,1.25-2.19-1.26-.72A4.42,4.42,0,0,0,4910.4,1492.83Z"
+              transform="translate(-4882.07 -1486.66)"
+            />
+          </svg>
+        </a>
+
+        <ul class="primary-submenu" aria-labelledby="setting_submenu">
+          <% settings_items.each do |settings_item| %>
+            <% if settings_item[:type] == :divider %>
+              <li class="divider-nav-item" aria-hidden="true"></li>
+            <% else %>
+              <li><%= menu_link(settings_item) %></li>
+            <% end %>
+          <% end %>
+        </ul>
+      </li>
+    <% else %>
+      <li class="primary-nav-item <%= item[:item_class] %>">
+        <%= menu_link(item) %>
+      </li>
+    <% end %>
+  <% end %>
+</ul>
