+++ app/components/page_block/navbar/wrapper/component.en.yml
@@ -1,0 +1,9 @@
+---
+en:
+  components:
+    page_block:
+      navbar:
+        wrapper:
+          menu: Menu
+          sign_up: Sign up
+          the_nonprofit_bike_registry: the non-profit bike registry
