+++ app/components/page_block/navbar/wrapper/component.html.erb
@@ -1,0 +1,38 @@
+<%# logo_only renders one link -- not worth a cache round trip %>
+<% cache_if(!@logo_only, cache_key) do %>
+  <nav class="primary-header-nav">
+    <div class="container">
+      <a href="<%= helpers.user_root_url %>" class="primary-logo">
+        <%= image_tag "revised/logo.svg", class: "primary-nav", alt: "Bike Index home" %>
+      </a>
+
+      <% unless @logo_only %>
+        <%= link_to news_path("bike-index--now-a-nonprofit"), class: "nonprofit-subtitle" do %>
+          <%= translation(".the_nonprofit_bike_registry") %>
+        <% end %>
+
+        <%= render organization_menu %>
+
+        <% if @current_user.blank? %>
+          <div class="center-navbar-signup-link-container">
+            <%= link_to translation(".sign_up"), new_user_url, class: "center-navbar-signup-link signup-link upcase" %>
+          </div>
+        <% end %>
+
+        <div id="menu-opened-backdrop" aria-hidden="true"></div>
+
+        <div class="hamburgler">
+          <button
+            id="primary_nav_hamburgler"
+            type="button"
+            aria-label="<%= translation(".menu") %>"
+            aria-expanded="false"
+            aria-controls="primary-main-menu"
+          ></button>
+        </div>
+
+        <%= render primary_menu %>
+      <% end %>
+    </div>
+  </nav>
+<% end %>
