+++ spec/components/page_block/navbar/primary_menu/component_spec.rb
@@ -2,26 +2,20 @@
 
 require "rails_helper"
 
-RSpec.describe PageBlock::Navbar::Component, type: :component do
-  it_behaves_like "cached_markup_digest"
-
+RSpec.describe PageBlock::Navbar::PrimaryMenu::Component, type: :component do
   let(:current_user) { nil }
-  let(:passive_organization) { nil }
-  let(:url) { "/" }
   let(:controller_namespace) { nil }
   let(:controller_name) { "welcome" }
   let(:action_name) { "index" }
   let(:instance) do
-    described_class.new(current_user:, current_user_or_unconfirmed_user: current_user, passive_organization:,
-      page_id: "welcome_index", controller_namespace:, controller_name:, action_name:)
+    described_class.new(current_user:, current_user_or_unconfirmed_user: current_user,
+      controller_namespace:, controller_name:, action_name:)
   end
   # The request drives active_link, which resolves the items that pass no :active
-  let(:component) { with_request_url(url) { render_inline(instance) } }
+  let(:component) { with_request_url("/") { render_inline(instance) } }
   let(:menu_links) { component.css("#primary-main-menu a").map { |link| link.text.strip } }
 
   it "renders the signed out menu" do
-    expect(component).to have_css "nav.primary-header-nav"
-    expect(component).to have_css "a.center-navbar-signup-link", text: "Sign up"
     expect(menu_links).to eq(["Search", "Marketplace", "Sign up", "log in", "Help",
       "Stolen bike?", "Donate", "Blog", "Marketplace", "Search"])
     expect(component).to_not have_css "#setting_submenu"
@@ -51,7 +45,6 @@ RSpec.describe PageBlock::Navbar::Component, type: :component do
 
     it "renders the settings submenu rather than the sign up links" do
       expect(component).to have_css "#setting_submenu"
-      expect(component).to_not have_css "a.center-navbar-signup-link"
       expect(component.css(".primary-submenu a").map { |link| link.text.strip })
         .to eq(["Your registrations", "Register a new bike", "party@bikeindex.org settings", "Logout"])
       expect(component).to have_css "#navUserSettingLink[data-email='party@bikeindex.org']"
@@ -66,45 +59,12 @@ RSpec.describe PageBlock::Navbar::Component, type: :component do
     end
 
     context "with an organization role" do
-      let(:organization) { FactoryBot.create(:organization, name: "Sweet Shop", short_name: "Sweet") }
+      let(:organization) { FactoryBot.create(:organization, name: "Sweet Shop") }
       let!(:organization_role) { FactoryBot.create(:organization_role_claimed, user: current_user, organization:) }
 
       it "links to the organization from the settings submenu" do
         expect(component).to have_css ".primary-submenu a", text: "View Sweet Shop"
       end
-
-      context "with a passive_organization" do
-        let(:passive_organization) { organization }
-
-        it "renders the organization submenu" do
-          expect(component).to have_css "#passive_organization_submenu", text: "Sweet"
-          expect(component).to have_css ".current-organization-submenu a"
-        end
-      end
-    end
-  end
-
-  context "with logo_only" do
-    let(:instance) { described_class.logo_only }
-
-    it "renders the logo without the menu" do
-      expect(component).to have_css "nav.primary-header-nav a.primary-logo"
-      expect(component).to_not have_css "#primary-main-menu"
-      expect(component).to_not have_css ".hamburgler"
-      expect(component).to_not have_css ".nonprofit-subtitle"
-    end
-  end
-
-  describe "caching", :caching do
-    include_context :caching_basic
-
-    # The cached fragment must include the locale in its key, or a request in
-    # one language serves the navbar cached in another. See ApplicationComponentHelper#cache.
-    it "varies the cached fragment by locale" do
-      en = with_request_url(url) { render_inline(instance) }.to_html
-      nl = I18n.with_locale(:nl) { with_request_url(url) { render_inline(instance) } }.to_html
-      expect(en).to include("Stolen bike?")
-      expect(nl).to_not include("Stolen bike?")
     end
   end
 end
